+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Changes since v18.04.0
 
 **Fixes:**
+- **Docker:** Fixed an issue where deleted qtrees could appear in Docker volume list.
 
 **Enhancements:**
 - **Kubernetes:** Updated etcd to v3.2.19 and client-go to v7.0.0.